fix(bots,backend): address Revi's gate findings (R1b58ea high + 2 low)
R1b58ea: under the shipped sandbox:auto a claw failure flattens to a
string at the __claw-runner IPC boundary and classifies UNCLASSIFIED —
which a FILTERED skip refuses by design, so plan_review_policy=skip
silently became wait on every sandboxed run (prod included). The bots'
skip route goes on:[any]: the operator's skip means the peer must NEVER
block, whatever killed it. The filtered-skip guard stays for
unsandboxed authors; the underlying engine class (typed classification
lost on the claw IPC wire, which also blinds the run-level usage-window
retry for sandboxed claw) is filed on the board (ADR-087 stage-3).

Rde458c: plan_review_policy gains [enum: wait, skip] — a typo'd --var
fails at launch instead of silently selecting wait. R84bcd0/Rae00df:
fillZeroValues emits float64 for int fields (the JSON-shaped contract
ValidateOutput and a store round-trip expect). Guard test updated to
pin on:[any] + the enum; ADR-091 second amendment records the
arbitration.

+++ pkg/backend/model/chain_skip_test.go
@@ -244,7 +244,9 @@ func TestFillZeroValues(t *testing.T) {
 	if out["concerns"] != "kept" {
 		t.Errorf("existing key overwritten: %+v", out)
 	}
-	if out["blocking"] != false || out["count"] != 0 || out["score"] != 0.0 {
+	// int fields synthesize float64 — the JSON-shaped contract
+	// ValidateOutput and a store round-trip both expect.
+	if out["blocking"] != false || out["count"] != float64(0) || out["score"] != 0.0 {
 		t.Errorf("zero values wrong: %+v", out)
 	}
 	if arr, ok := out["items"].([]any); !ok || len(arr) != 0 {
